some very naive knn retrieval method.
performs O(n) operations, and then sorts the result. At least the
sorting might be partial :-).

This is the groundwork for a very basic knn classifier, and performance
is currently of no concern.

+++ src/ML/LinearRegression.hs
@@ -172,7 +172,7 @@ fStatistics m =
 mse :: LinearModel a => a -> Double
 mse m =
     let respScale = snd $ scalingFactor $ response $ M.modelSpec' m
-    in rss m * respScale^2 / degreesOfFredom m
+    in rss m * respScale^(2::Int) / degreesOfFredom m
 
 rse :: LinearModel a => a -> Double
 rse = sqrt . mse
